viewing other people's videos is working

+++ Activity 4/flask/app.py
@@ -41,6 +41,20 @@ app.config['CORS_HEADERS'] = 'Content-Type'
 
 cors = CORS(app)
 
+testuser2 = 'admin2'
+testuser2hashedpass = generate_password_hash('admin2')
+cursor.execute("INSERT INTO users(Username, EncryptedPass, TotalVideoCount, \
+                DateCreated) VALUES ('{}', '{}', 0, '{}')".format(testuser2, \
+                testuser2hashedpass, datetime.datetime.now().strftime('%Y-%m-%d')))
+testuser1 = 'admin'
+testuser1hashedpass = generate_password_hash('admin')
+cursor.execute("INSERT INTO users(Username, EncryptedPass, TotalVideoCount, \
+                DateCreated) VALUES ('{}', '{}', 0, '{}')".format(testuser1, \
+                testuser1hashedpass, datetime.datetime.now().strftime('%Y-%m-%d')))                
+cursor.close()
+conn.commit()
+conn.close()
+
 
 @app.route("/")
 def home():
@@ -50,28 +64,23 @@ def home():
 @app.route("/login", methods=['GET','POST'])
 @limiter.limit("14400/day;600/hour;10/minute")
 def login():
-
+    
+    conn = pymysql.connect('mysql', 'root', 'root', 'db')
+    cursor = conn.cursor()
     if request.method == 'GET':
         return redirect('http://localhost:5000')
-    testuser1 = 'admin'
-    testuser1hashedpass = generate_password_hash('admin')
     username = request.form['username']
     password = request.form['password']
     hashedpass = generate_password_hash(password)
-    cursor.execute("INSERT INTO users(Username, EncryptedPass, TotalVideoCount, \
-                    DateCreated) VALUES ('{}', '{}', 0, '{}')".format(testuser1, \
-                    testuser1hashedpass, datetime.datetime.now().strftime('%Y-%m-%d')))
     cursor.execute("SELECT EncryptedPass FROM users WHERE Username='{}'".format(str(username)))
     userpass = cursor.fetchone()
-	
+    cursor.close()
+    conn.close()
     if userpass == None:
-        conn.commit()
         return render_template('wronguser.html')
     elif check_password_hash(userpass[0], password):
         session['username'] = username
-        conn.commit()
         return redirect(url_for('homepage'))
-    conn.commit()
     return redirect(url_for('wrongpass'))
 
 @app.route("/logout", methods=['GET','POST'])
@@ -94,7 +103,9 @@ def wrongpass():
 
 @app.route("/homepage", methods=['GET','POST'])
 def homepage():
-    
+
+    conn = pymysql.connect('mysql', 'root', 'root', 'db')
+    cursor = conn.cursor()
     if 'username' in session:
         if request.method == 'POST':
             target = os.path.join(APP_ROOT, "static")
@@ -115,12 +126,14 @@ def homepage():
                     cursor.execute("SELECT UserID FROM users WHERE Username='{}'".format((session['username'])))
                     userid = cursor.fetchone()
                     print(userid, file=sys.stderr)
-                    cursor.execute("INSERT INTO video(UserID, VideoTitle, VideoURL, DateUploaded) VALUES \
-                        ('{}', '{}', '{}', '{}')".format(userid[0], localfile, str(destination),\
-                        datetime.datetime.now().strftime('%Y-%m-%d')))
+                    cursor.execute("INSERT INTO video(UserID, VideoTitle, VideoUser, VideoURL, DateUploaded) VALUES \
+                        ('{}', '{}', '{}', '{}', '{}')".format(userid[0], localfile, str(destination),\
+                        session['username'], datetime.datetime.now().strftime('%Y-%m-%d')))
                     cursor.execute("UPDATE users SET TotalVideoCount = TotalVideoCount + \
                         1 WHERE Username = '{}'".format(str(session['username'])))
                     conn.commit()
+                    cursor.close()
+                    conn.close()
                     return render_template('homepage.html', username = session['username'])
             else:
                 for f in request.files.getlist("file"):
@@ -133,20 +146,29 @@ def homepage():
                     f.save(destination)
                     cursor.execute("SELECT UserID FROM users WHERE Username='{}'".format((session['username'])))
                     userid = cursor.fetchone()
-                    cursor.execute("INSERT INTO video(UserID, VideoTitle, VideoURL, DateUploaded) VALUES \
-                        ('{}', '{}', '{}', '{}')".format(userid[0], filename, \
-                        str(destination), datetime.datetime.now().strftime('%Y-%m-%d')))
+                    cursor.execute("INSERT INTO video(UserID, VideoTitle, VideoURL, VideoUser, DateUploaded) VALUES \
+                        ('{}', '{}', '{}', '{}', '{}')".format(userid[0], filename, \
+                        str(destination), session['username'], datetime.datetime.now().strftime('%Y-%m-%d')))
                     cursor.execute("UPDATE users SET TotalVideoCount = TotalVideoCount + \
                         1 WHERE Username = '{}'".format(str(session['username'])))
                     conn.commit()
+                    cursor.close()
+                    conn.close()
                     return render_template('homepage.html', username = session['username'])
+        cursor.close()
+        conn.close()
         return render_template('homepage.html', username = session['username'])
     else:
+        cursor.close()
+        conn.close()
         return redirect(url_for('login'))
 
 
 @app.route('/getvideos', methods=['GET', 'POST'])
 def getvideos():
+
+    conn = pymysql.connect('mysql', 'root', 'root', 'db')
+    cursor = conn.cursor()
     if 'username' in session:
         username = request.get_json()
         username = username['username']
@@ -160,8 +182,37 @@ def getvideos():
         for result in rows:
             json_data.append(dict(zip(row_headers,result)))
         print(json_data, file=sys.stderr)
+        cursor.close()
+        conn.close()
         return jsonify(json_data)
+    cursor.close()
+    conn.close()
+    return redirect(url_for('login'))
 
+@app.route('/getothervids', methods=['GET', 'POST'])
+def getothervids():
+
+    conn = pymysql.connect('mysql', 'root', 'root', 'db')
+    cursor = conn.cursor()
+    if 'username' in session:
+        
+        username = request.get_json()
+        username = username['username']
+        cursor.execute("SELECT UserID FROM users WHERE Username='{}'".format(username))
+        userid = cursor.fetchone()
+        cursor.execute("SELECT * FROM video WHERE UserID!={}".format(userid[0]))
+        rows = cursor.fetchall()
+        row_headers=[x[0] for x in cursor.description]
+        json_data=[]
+        for result in rows:
+            json_data.append(dict(zip(row_headers,result)))
+        print(json_data, file=sys.stderr)
+        cursor.close()
+        conn.close()
+        return jsonify(json_data)
+    cursor.close()
+    conn.close()
+    return redirect(url_for('login'))
 
 @app.route('/videos/<title>')
 def videos(title):
@@ -171,12 +222,17 @@ def videos(title):
 
 @app.route('/delete/<videoid>')
 def delete(videoid):
+
+    conn = pymysql.connect('mysql', 'root', 'root', 'db')
+    cursor = conn.cursor()
     print(videoid, file=sys.stderr)
     if 'username' in session:
         cursor.execute("SELECT VideoTitle FROM video WHERE VideoID={}".format(videoid))
         tempFile = cursor.fetchone()
-        print(tempFile, file=sys.stderr)
         tempFile = tempFile[0]
+        print(tempFile, file=sys.stderr)
+        if tempFile == '':
+            return redirect(url_for('homepage'))
         cursor.execute("DELETE FROM video WHERE VideoID={}".format(videoid))
         cursor.execute("SELECT UserID FROM users WHERE Username='{}'".format((session['username'])))
         userid = cursor.fetchone()
@@ -184,7 +240,11 @@ def delete(videoid):
                     1 WHERE Username = '{}'".format(str(session['username'])))
         conn.commit()
         os.remove("static/"+tempFile)
+        cursor.close()
+        conn.close()
         return redirect(url_for('homepage'))
+    cursor.close()
+    conn.close()
     return redirect(url_for('login'))
